+++ packages/@aws-cdk/toolkit-lib/lib/toolkit/toolkit.ts
@@ -11,8 +11,7 @@ import { AssetBuildTime, type DeployOptions } from '../actions/deploy';
 import { type ExtendedDeployOptions, buildParameterMap, createHotswapPropertyOverrides, removePublishedAssets } from '../actions/deploy/private';
 import { type DestroyOptions } from '../actions/destroy';
 import type { DiffOptions } from '../actions/diff';
-import { makeTemplateInfos } from '../actions/diff';
-import { determinePermissionType } from '../actions/diff/private';
+import { determinePermissionType, makeTemplateInfos } from '../actions/diff/private';
 import { type ListOptions } from '../actions/list';
 import { type RollbackOptions } from '../actions/rollback';
 import { type SynthOptions } from '../actions/synth';
